#2 move Extent to visualise.py
